Renamed logistic regression model
